controller: remove IPAMClaim finalizers on VMI delete events
This way, we ensure the CNI only returns the IP addresses to the pool
after the VMI has been deleted on VM foreground deletes.

Without it, the IPAMClaim would be deleted by the Kubernetes garbage
collector before the VMI was deleted, which would cause the IP address
to be released to the pool (by the CNI) prematurely. More details on
[0].

[0] - https://kubernetes.io/docs/concepts/architecture/garbage-collection/#foreground-deletion

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>