claims: Filter ipamclaim lookup according namespace
Currently looking for matching ipam claims is done only
according the owning VM name.
It means that when removing finalizer, we will remove
the finalizer from ipam claims with the same owning VM name across
all namespaces, which is wrong.
Fix it by limiting the search to the specific VM namespace.

Signed-off-by: Or Shoval <oshoval@redhat.com>