claims: Retry in case the claim get failed
In case we try to create a claim, and it already exists,
and then we try to fetch it in order to validate it,
but it was deleted meanwhile, we should retry to reconcile,
in order to give it another chance to be created.

Also no need double check err != nil, as we check it already in this block.

Signed-off-by: Or Shoval <oshoval@redhat.com>