tests/e2e, persistentip: Move GetIPs to general function
In the current tests (where we only run persistent ip tests for
secondary interfaces), the ips of the VMI is extracted via vmi.status.
This is done both directly and via the matchers in testenv library.

In order for the test suite to support running the same tests on primary
UDN VMs, this function needs to be generalized in the following ways:
1. it should align all the IP retrievals to use this new function (right
now the IPs are retrieved both directly accessing vmi.status and via a
function).
2. it should move the function retrieving the IPs to a first-class
functions variable, so that the way IPS are retrieved can be
generalized.
3. it should only have the common parameters needed to get the IPs
(which is the VMI object).

Moving get the IP extraction to a general function, so that it could be
consumed uniformly. Doing this will allow to more easily retrieve the IP
in other cases such as primary UDN, where the IP is currently retrieved
in another way. This will be added in future commits.

Signed-off-by: Ram Lavi <ralavi@redhat.com>