tests/e2e, persistentip: Move generateVMI to general function
In the current tests (where we only run persistent ip tests for
secondary interfaces), the VMI is created using the
GenerateAlpineWithMultusVMI function.

In order for the test suite to support running the same tests on primary
UDN VMIs, refactoring the function to use VMI factory function.

Moreover, the NAD Name is changed to be static, as it allows it to be
consumed before runtime, as a test parameter for both the NAD and VMI
creation. This shouldn't interfere with the tests operations as the NADs
are created in different namespaces.

Signed-off-by: Ram Lavi <ralavi@redhat.com>