tests/e2e, persistentip: Introduce primary-UDN helper functions
Introducing the functions needed in order to run the current test suite
for primary UDN interfaces:
- defaultNetworkStatusAnnotationIPs - as the getIP function for primary
UDN.
- vmiWithPasst - as the vmi Generating function for primary UDN.
These will be introduced as parameter functions when primary-UDN Entry
is introduced in future commit.

Moreover, since the function getting the IPs for primary-UDN case needs
to return error, this is added to the general function signature.

Signed-off-by: Ram Lavi <ralavi@redhat.com>