e2e, persistent-ips: Get primary UDN ips from vmi status
Due to improvements in kubevirt/kubevirt the interface on vmi.status [0]
the vmi interface is now published on the vmi.status. This allows
retrieving the IPs of the primary UDN interface via the vmi.status.
Changing the primary UDN e2e function that retrieves the IPs to use the
vmi.status.

[0] https://github.com/kubevirt/kubevirt/pull/13018

Signed-off-by: Ram Lavi <ralavi@redhat.com>