nad: Skip NADs with empty config
Until now, a NAD with empty config, was considered invalid,
causing an error / rejecting it on the webhook.
Istio creates such NADs, so we can't consider them as fatal,
once we list all the NADs in a namespace.
Otherwise the webhook will block VM creation in case it uses
such NAD.

Therefore just ignore those.

Signed-off-by: Or Shoval <oshoval@redhat.com>