config: Provide network-policies manifests
In a scenario where the cluster networking has been hardened and futures
a default deny-all network policy that affecting kubevirt-ipam-controller
namespace, it will not be able to operate properly.

Provide network-policies for the project in order to comply with such
restrictions in form of deny-all network-policy.
The introduced network-policies should allow kubevirt-ipam-controller
operate when a deny-all network-policy is in place.

This commit add the following network-polices:
1. Allow egress to cluster DNS
   Enable interaction with cluster DNS, w/o this policy the
   kubevirt-ipam-controller's manager cannot establish connection with
   the cluster API.
   Affects pods with app=ipam-virt-workloads label.
2. Allow egress to kube-apiserver
   Enable interaction with the cluster API, w/o this policy the
   controller cannot r/w VMs/Pods/IPAMCliams, etc..
   Affects pods with with app=ipam-virt-workloads label.
3. Allow ingress to kubevirt-ipam-controller webhook-service
   Allow pod in other namespaces interact with the webhook.
   w/o this policy VM creation fail with context-deadline error due to
   blocked connection.
   Affects pods with app=ipam-virt-workloads label.

This change include re-generated yaml files under config, following 'make
build-installer'.

Testing the network-policies will be done in follow-up commit.

Signed-off-by: Or Mergi <ormergi@redhat.com>